fix(finance): the rate repair reaches invoices already in the mirror
The writer fix alone could not repair anything. The sync skips an
invoice whose SOURCE values are unchanged, and the source of an already
mirrored invoice does not change just because we started recording a
rate — so every row written before the fix would keep a null rate for
good, one unconvertible row refuses the whole total (FIN-AC-6), and a
deployed installation would show an empty finance card forever.

Two halves, and both are needed.

Migration 0205 repairs the rows already in the database. Only the
identity, matching the writer exactly: a foreign invoice still gets no
rate, because inventing one would sum dollars into euros silently.

The sync's skip now asks whether the row carries a rate it SHOULD have.
The hash covers the source's values, and the rate is not one of them —
it is derived from the workspace's base currency, which the source knows
nothing about. Without this a future base-currency change reopens the
same hole, and the migration would have to be written again.

An integration test proves both directions: a rate-stripped ledger is
repaired on the next pass, and the pass after that writes nothing. A
repair that re-fired every pass is the same bug with the opposite sign.

Mutation-checked: forcing the skip fails it.

mirror.go was over the 500-line cap, so payments moved to their own
file — they were always a separate concept in it.
